requirements.txt
alembic==1.8.1
amqp==5.1.1
attrs==22.1.0
bcrypt==4.0.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
distro==1.7.0
docker==6.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
Flask==2.2.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
kombu==5.2.4
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
paramiko==2.11.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==5.4.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
texttable==1.6.4
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
Werkzeug==2.2.1
zipp==3.8.1
